+++ psakefile.ps1
@@ -8,19 +8,25 @@ Properties {
         $ProjectRoot = $PSScriptRoot
     }
 
-    # Determine the folder names for staging the module
-    $StagingFolder = Join-Path -Path $ProjectRoot -ChildPath 'staging'
-    $ModuleFolder = Join-Path -Path $StagingFolder -ChildPath 'CosmosDB'
-
     $Timestamp = Get-Date -uformat "%Y%m%d-%H%M%S"
     $PSVersion = $PSVersionTable.PSVersion.Major
     $separator = '----------------------------------------------------------------------'
 }
 
-Task Default -Depends Build
+Task Default -Depends Test,Build
 
 Task Init {
     Set-Location -Path $ProjectRoot
+
+    # Install any dependencies required for the Init stage
+    Invoke-PSDepend `
+        -Path $PSScriptRoot `
+        -Force `
+        -Import `
+        -Install `
+        -Tags 'Init'
+
+    Set-BuildEnvironment -Force
 
     $separator
     'Build System Details:'
@@ -41,15 +47,13 @@ Task Init {
 Task Test -Depends Init {
     $separator
 
-    # Prepare test environment by ensuring appropriate version of AzureRM modules are installed
-    if ($PSVersionTable.PSEdition -eq 'Core')
-    {
-        Install-ModuleMultiScoped -Name AzureRM.NetCore
-    }
-    else
-    {
-        Install-ModuleMultiScoped -Name AzureRM
-    }
+    # Install any dependencies required for the Test stage
+    Invoke-PSDepend `
+        -Path $PSScriptRoot `
+        -Force `
+        -Import `
+        -Install `
+        -Tags 'Test',('Test_{0}' -f $PSVersionTable.PSEdition)
 
     # Execute tests
     $testResultsFile = Join-Path -Path $ProjectRoot -ChildPath 'test\TestResults.xml'
@@ -119,12 +123,20 @@ Task Test -Depends Init {
     "`n"
 }
 
-Task Build -Depends Test {
+Task Build -Depends Init {
     $separator
+
+    # Install any dependencies required for the Build stage
+    Invoke-PSDepend `
+        -Path $PSScriptRoot `
+        -Force `
+        -Import `
+        -Install `
+        -Tags 'Build'
 
     # Generate the next version by adding the build system build number to the manifest version
     $manifestPath = Join-Path -Path $ProjectRoot -ChildPath 'src/CosmosDB.psd1'
-    $newVersion = New-VersionNumber `
+    $newVersion = Get-VersionNumber `
         -ManifestPath $manifestPath `
         -Build $ENV:BHBuildNumber
 
@@ -135,36 +147,40 @@ Task Build -Depends Test {
     }
 
     # Determine the folder names for staging the module
-    $VersionFolder = Join-Path -Path $ModuleFolder -ChildPath $newVersion
+    $StagingFolder = Join-Path -Path $ProjectRoot -ChildPath 'staging'
+    $ModuleFolder = Join-Path -Path $StagingFolder -ChildPath 'CosmosDB'
+
+    # Determine the folder names for staging the module
+    $versionFolder = Join-Path -Path $ModuleFolder -ChildPath $newVersion
 
     # Stage the module
     $null = New-Item -Path $StagingFolder -Type directory -ErrorAction SilentlyContinue
     $null = New-Item -Path $ModuleFolder -Type directory -ErrorAction SilentlyContinue
-    Remove-Item -Path $VersionFolder -Recurse -Force -ErrorAction SilentlyContinue
-    $null = New-Item -Path $VersionFolder -Type directory
+    Remove-Item -Path $versionFolder -Recurse -Force -ErrorAction SilentlyContinue
+    $null = New-Item -Path $versionFolder -Type directory
 
     # Populate Version Folder
-    $null = Copy-Item -Path (Join-Path -Path $ProjectRoot -ChildPath 'src/CosmosDB.psd1') -Destination $VersionFolder
-    $null = Copy-Item -Path (Join-Path -Path $ProjectRoot -ChildPath 'src/CosmosDB.psm1') -Destination $VersionFolder
-    $null = Copy-Item -Path (Join-Path -Path $ProjectRoot -ChildPath 'src/lib') -Destination $VersionFolder -Recurse
-    $null = Copy-Item -Path (Join-Path -Path $ProjectRoot -ChildPath 'src/formats') -Destination $VersionFolder -Recurse
-    $null = Copy-Item -Path (Join-Path -Path $ProjectRoot -ChildPath 'src/types') -Destination $VersionFolder -Recurse
-    $null = Copy-Item -Path (Join-Path -Path $ProjectRoot -ChildPath 'src/en-US') -Destination $VersionFolder -Recurse
-    $null = Copy-Item -Path (Join-Path -Path $ProjectRoot -ChildPath 'LICENSE') -Destination $VersionFolder
-    $null = Copy-Item -Path (Join-Path -Path $ProjectRoot -ChildPath 'README.md') -Destination $VersionFolder
-    $null = Copy-Item -Path (Join-Path -Path $ProjectRoot -ChildPath 'CHANGELOG.md') -Destination $VersionFolder
-    $null = Copy-Item -Path (Join-Path -Path $ProjectRoot -ChildPath 'RELEASENOTES.md') -Destination $VersionFolder
+    $null = Copy-Item -Path (Join-Path -Path $ProjectRoot -ChildPath 'src/CosmosDB.psd1') -Destination $versionFolder
+    $null = Copy-Item -Path (Join-Path -Path $ProjectRoot -ChildPath 'src/CosmosDB.psm1') -Destination $versionFolder
+    $null = Copy-Item -Path (Join-Path -Path $ProjectRoot -ChildPath 'src/lib') -Destination $versionFolder -Recurse
+    $null = Copy-Item -Path (Join-Path -Path $ProjectRoot -ChildPath 'src/formats') -Destination $versionFolder -Recurse
+    $null = Copy-Item -Path (Join-Path -Path $ProjectRoot -ChildPath 'src/types') -Destination $versionFolder -Recurse
+    $null = Copy-Item -Path (Join-Path -Path $ProjectRoot -ChildPath 'src/en-US') -Destination $versionFolder -Recurse
+    $null = Copy-Item -Path (Join-Path -Path $ProjectRoot -ChildPath 'LICENSE') -Destination $versionFolder
+    $null = Copy-Item -Path (Join-Path -Path $ProjectRoot -ChildPath 'README.md') -Destination $versionFolder
+    $null = Copy-Item -Path (Join-Path -Path $ProjectRoot -ChildPath 'CHANGELOG.md') -Destination $versionFolder
+    $null = Copy-Item -Path (Join-Path -Path $ProjectRoot -ChildPath 'RELEASENOTES.md') -Destination $versionFolder
 
     # Prepare external help
     'Building external help file'
     New-ExternalHelp `
         -Path (Join-Path -Path $ProjectRoot -ChildPath 'docs\') `
-        -OutputPath $VersionFolder `
+        -OutputPath $versionFolder `
         -Force
 
     # Set the new version number in the staged Module Manifest
     'Updating module manifest'
-    $stagedManifestPath = Join-Path -Path $VersionFolder -ChildPath 'CosmosDB.psd1'
+    $stagedManifestPath = Join-Path -Path $versionFolder -ChildPath 'CosmosDB.psd1'
     $stagedManifestContent = Get-Content -Path $stagedManifestPath -Raw
     $stagedManifestContent = $stagedManifestContent -replace '(?<=ModuleVersion\s+=\s+'')(?<ModuleVersion>.*)(?='')', $newVersion
     $stagedManifestContent = $stagedManifestContent -replace '## What is New in CosmosDB Unreleased', "## What is New in CosmosDB $newVersion"
@@ -172,123 +188,72 @@ Task Build -Depends Test {
 
     # Set the new version number in the staged CHANGELOG.md
     'Updating CHANGELOG.MD'
-    $stagedChangeLogPath = Join-Path -Path $VersionFolder -ChildPath 'CHANGELOG.md'
+    $stagedChangeLogPath = Join-Path -Path $versionFolder -ChildPath 'CHANGELOG.md'
     $stagedChangeLogContent = Get-Content -Path $stagedChangeLogPath -Raw
     $stagedChangeLogContent = $stagedChangeLogContent -replace '# Unreleased', "# $newVersion"
     Set-Content -Path $stagedChangeLogPath -Value $stagedChangeLogContent -NoNewLine -Force
 
     # Set the new version number in the staged RELEASENOTES.md
     'Updating RELEASENOTES.MD'
-    $stagedReleaseNotesPath = Join-Path -Path $VersionFolder -ChildPath 'RELEASENOTES.md'
+    $stagedReleaseNotesPath = Join-Path -Path $versionFolder -ChildPath 'RELEASENOTES.md'
     $stagedReleaseNotesContent = Get-Content -Path $stagedReleaseNotesPath -Raw
     $stagedReleaseNotesContent = $stagedReleaseNotesContent -replace '## What is New in CosmosDB Unreleased', "## What is New in CosmosDB $newVersion"
     Set-Content -Path $stagedReleaseNotesPath -Value $stagedReleaseNotesContent -NoNewLine -Force
 
-    "`n"
-}
-
-Task Deploy -Depends Build {
-    $separator
-
-    # Generate the next version by adding the build system build number to the manifest version
-    $manifestPath = Join-Path -Path $ProjectRoot -ChildPath 'src/CosmosDB.psd1'
-    $newVersion = New-VersionNumber `
-        -ManifestPath $manifestPath `
-        -Build $ENV:BHBuildNumber
-
-    # Determine the folder names for staging the module
-    $VersionFolder = Join-Path -Path $ModuleFolder -ChildPath $newVersion
-
-    # Copy the module to the PSModulePath
-    $PSModulePath = ($ENV:PSModulePath -split ';')[0]
-
-    "Copying Module to $PSModulePath"
-    Copy-Item `
-        -Path $ModuleFolder `
-        -Destination $PSModulePath `
-        -Recurse `
-        -Force
-
     # Create zip artifact
-    $zipFilePath = Join-Path `
+    $zipFileFolder = Join-Path `
         -Path $StagingFolder `
-        -ChildPath "${ENV:BHProjectName}_${ENV:BHBuildNumber}.zip"
+        -ChildPath 'zip'
+
+    $null = New-Item -Path $zipFileFolder -Type directory -ErrorAction SilentlyContinue
+
+    $zipFilePath = Join-Path `
+        -Path $zipFileFolder `
+        -ChildPath "${ENV:BHProjectName}_$newVersion.zip"
     $null = Add-Type -assemblyname System.IO.Compression.FileSystem
     [System.IO.Compression.ZipFile]::CreateFromDirectory($ModuleFolder, $zipFilePath)
 
-    if ($ENV:BHBuildSystem -eq 'AppVeyor')
+    # Update the Git Repo if this is the master branch build in VSTS
+    if ($ENV:BHBuildSystem -eq 'VSTS')
     {
-        # If AppVeyor, publish the deploy artefacts for debug purposes
-        "Pushing package $zipFilePath as Appveyor artifact"
-        Push-AppveyorArtifact $zipFilePath
-        Remove-Item -Path $zipFilePath -Force
-
-        <#
-            If this is a build of the Master branch and not a PR push
-            then publish the Module to the PowerShell Gallery.
-        #>
         if ($ENV:BHBranchName -eq 'master')
         {
-            $commitMessage = $ENV:BHCommitMessage.TrimEnd()
-            "Commit to Master branch detected with commit message: '$commitMessage'"
+            # This is a push to master so update GitHub with release info
+            'Beginning update to master branch with deployed information'
 
-            if ($commitMessage -match ' Deploy!$')
+            $commitMessage = $ENV:BHCommitMessage.TrimEnd()
+            "Commit to master branch triggered with commit message: '$commitMessage'"
+
+            if ($commitMessage -match '^Azure DevOps Deploy updating Version Number to [0-9/.]*')
             {
                 # This was a deploy commit so no need to do anything
-                'Skipping deployment because this was a commit triggered by a deployment'
-            }
-            elseif ($ENV:APPVEYOR_PULL_REQUEST_NUMBER)
-            {
-                # This is a PR so do nothing
-                'Skipping deployment because this is a Pull Request'
+                'Skipping update to master branch with deployed information because this was triggered by Azure DevOps Updating the Version Number'
             }
             else
             {
-                # This is a commit to Master
-                'Publishing Module to PowerShell Gallery'
-                Get-PackageProvider `
-                    -Name NuGet `
-                    -ForceBootstrap
-                Publish-Module `
-                    -Name 'CosmosDB' `
-                    -RequiredVersion $newVersion `
-                    -NuGetApiKey $ENV:PowerShellGalleryApiKey `
-                    -Confirm:$false
-
-                # This is not a PR so deploy
-                'Beginning update to master branch with deployed information'
-
                 # Pull the master branch, update the readme.md and manifest
                 Set-Location -Path $ProjectRoot
                 Invoke-Git -GitParameters @('config', '--global', 'credential.helper', 'store')
 
-                Add-Content `
-                    -Path "$env:USERPROFILE\.git-credentials" `
-                    -Value "https://$($env:GitHubPushFromPlagueHO):x-oauth-basic@github.com`n"
-
-                Invoke-Git -GitParameters @('config', '--global', 'user.email', 'plagueho@gmail.com')
-                Invoke-Git -GitParameters @('config', '--global', 'user.name', 'Daniel Scott-Raynsford')
-                Invoke-Git -GitParameters @('checkout', '-f', 'master')
-
                 # Replace the manifest with the one that was published
-                'Updating files changed during deployment.='
+                'Updating files changed during deployment'
                 Copy-Item `
-                    -Path (Join-Path -Path $VersionFolder -ChildPath 'CosmosDB.psd1') `
+                    -Path $stagedManifestPath `
                     -Destination (Join-Path -Path $ProjectRoot -ChildPath 'src') `
                     -Force
                 Copy-Item `
-                    -Path (Join-Path -Path $VersionFolder -ChildPath 'CHANGELOG.MD') `
+                    -Path $stagedChangeLogPath `
                     -Destination $ProjectRoot `
                     -Force
                 Copy-Item `
-                    -Path (Join-Path -Path $VersionFolder -ChildPath 'RELEASENOTES.MD') `
+                    -Path $stagedReleaseNotesPath `
                     -Destination $ProjectRoot `
                     -Force
 
                 # Update the master branch
                 'Pushing deployment changes to Master'
                 Invoke-Git -GitParameters @('add', '.')
-                Invoke-Git -GitParameters @('commit', '-m', "$NewVersion Deploy!")
+                Invoke-Git -GitParameters @('commit', '-m', "Azure DevOps Deploy updating Version Number to $NewVersion")
                 Invoke-Git -GitParameters @('status')
                 Invoke-Git -GitParameters @('push', 'origin', 'master')
 
@@ -304,14 +269,73 @@ Task Deploy -Depends Build {
                 Invoke-Git -GitParameters @('push', 'origin', 'dev')
             }
         }
+        else
+        {
+            "Skipping update to master branch with deployed information because branch is: '$ENV:BHBranchName'"
+        }
     }
+    else
+    {
+        "Skipping update to master branch with deployed information because build system is: '$ENV:BHBuildSystem'"
+    }
+    "`n"
+}
+
+Task Deploy {
+    $separator
+
+    # Determine the folder name for the Module
+    $ModuleFolder = Join-Path -Path $ProjectRoot -ChildPath 'CosmosDB'
+
+    # Install any dependencies required for the Deploy stage
+    Invoke-PSDepend `
+        -Path $PSScriptRoot `
+        -Force `
+        -Import `
+        -Install `
+        -Tags 'Deploy'
+
+    # Copy the module to the PSModulePath
+    $PSModulePath = ($ENV:PSModulePath -split ';')[0]
+    $destinationPath = Join-Path -Path $PSModulePath -ChildPath 'CosmosDB'
+
+    "Copying Module from $ModuleFolder to $destinationPath"
+    Copy-Item `
+        -Path $ModuleFolder `
+        -Destination $destinationPath `
+        -Container `
+        -Recurse `
+        -Force
+
+    $installedModule = Get-Module -Name CosmosDB -ListAvailable
+
+    $versionNumber = $installedModule.Version |
+        Sort-Object -Descending |
+        Select-Object -First 1
+
+    if (-not $versionNumber)
+    {
+        Throw "CosmosDB Module could not be found after copying to $PSModulePath"
+    }
+
+    # This is a deploy from the staging folder
+    "Publishing CosmosDB Module version '$versionNumber' to PowerShell Gallery"
+    $null = Get-PackageProvider `
+        -Name NuGet `
+        -ForceBootstrap
+
+    Publish-Module `
+        -Name 'CosmosDB' `
+        -RequiredVersion $versionNumber `
+        -NuGetApiKey $ENV:PowerShellGalleryApiKey `
+        -Confirm:$false
 }
 
 <#
     .SYNOPSIS
         Generate a new version number.
 #>
-function New-VersionNumber
+function Get-VersionNumber
 {
     param
     (
@@ -371,36 +395,5 @@ function Invoke-Git
     catch
     {
         Write-Warning -Message $_
-    }
-}
-
-<#
-    .SYNOPSIS
-        Try to install a module to the machine scope, but try the
-        CurrentUser scope if not allowed.
-#>
-function Install-ModuleMultiScoped
-{
-    param
-    (
-        [Parameter(Mandatory = $true)]
-        [System.String]
-        $Name
-    )
-
-    $installModuleParameters = @{
-        Name = $Name
-        Force = $true
-        AllowClobber = $true
-        Repository = 'PSGallery'
-    }
-
-    try
-    {
-        Install-Module @installModuleParameters
-    }
-    catch
-    {
-        Install-Module @installModuleParameters -Scope CurrentUser
     }
 }